Export mockGlobal, reorganize utils
