+++ src/index.js
@@ -1,1 +1,2 @@
 export {createTestDriver} from './screen-driver';
+export {mockGlobal} from './utils/mockGlobal';
